add users role (reader, manager, administrator)

+++ src/bot/plugins/common.py
@@ -4,32 +4,44 @@ from pyrogram.types import InlineKeyboardButton, InlineKeyboardMarkup
 
 from ... import db_management
 from ...qbittorrent_manager import QbittorrentManagement
+from ...utils import get_user_from_config
+from ...configs.enums import UserRolesEnum
 
 
-async def send_menu(client: Client, message, chat) -> None:
-    db_management.write_support("None", chat)
+async def send_menu(client: Client, message_id: int, chat_id: int) -> None:
+    user = get_user_from_config(chat_id)
     buttons = [
-        [InlineKeyboardButton("📝 List", "list")],
-        [InlineKeyboardButton("➕ Add Magnet", "category#add_magnet"),
-         InlineKeyboardButton("➕ Add Torrent", "category#add_torrent")],
-        [InlineKeyboardButton("⏸ Pause", "pause"),
-         InlineKeyboardButton("▶️ Resume", "resume")],
-        [InlineKeyboardButton("⏸ Pause All", "pause_all"),
-         InlineKeyboardButton("▶️ Resume All", "resume_all")],
-        [InlineKeyboardButton("🗑 Delete", "delete_one"),
-         InlineKeyboardButton("🗑 Delete All", "delete_all")],
-        [InlineKeyboardButton("➕ Add Category", "add_category"),
-         InlineKeyboardButton("🗑 Remove Category", "select_category#remove_category")],
-        [InlineKeyboardButton("📝 Modify Category", "select_category#modify_category")],
-        [InlineKeyboardButton("⚙️ Settings", "settings")]
+        [InlineKeyboardButton("📝 List", "list")]
     ]
 
+    if user.role in [UserRolesEnum.Manager, UserRolesEnum.Administrator]:
+        buttons += [
+            [InlineKeyboardButton("➕ Add Magnet", "category#add_magnet"),
+             InlineKeyboardButton("➕ Add Torrent", "category#add_torrent")],
+            [InlineKeyboardButton("⏸ Pause", "pause"),
+             InlineKeyboardButton("▶️ Resume", "resume")],
+            [InlineKeyboardButton("⏸ Pause All", "pause_all"),
+             InlineKeyboardButton("▶️ Resume All", "resume_all")],
+        ]
+
+    if user.role == UserRolesEnum.Administrator:
+        buttons += [
+            [InlineKeyboardButton("🗑 Delete", "delete_one"),
+             InlineKeyboardButton("🗑 Delete All", "delete_all")],
+            [InlineKeyboardButton("➕ Add Category", "add_category"),
+             InlineKeyboardButton("🗑 Remove Category", "select_category#remove_category")],
+            [InlineKeyboardButton("📝 Modify Category", "select_category#modify_category")],
+            [InlineKeyboardButton("⚙️ Settings", "settings")]
+        ]
+
+    db_management.write_support("None", chat_id)
+
     try:
-        await client.edit_message_text(chat, message, text="Qbittorrent Control",
+        await client.edit_message_text(chat_id, message_id, text="Qbittorrent Control",
                                        reply_markup=InlineKeyboardMarkup(buttons))
 
     except MessageIdInvalid:
-        await client.send_message(chat, text="Qbittorrent Control", reply_markup=InlineKeyboardMarkup(buttons))
+        await client.send_message(chat_id, text="Qbittorrent Control", reply_markup=InlineKeyboardMarkup(buttons))
 
 
 async def list_active_torrents(client: Client, n, chat, message, callback, status_filter: str = None) -> None:
